begun exceeding calories logic

+++ src/main/goal/FitnessGoal.cs
@@ -12,7 +12,7 @@ public class FitnessGoal : GoalDecorator
         this.RecommendedWorkouts = recommendedWorkouts;
     }
 
-    public int GetAdditionalCalories()
+    public double GetAdditionalCalories()
     {
         return -1;
     }
