+++ src/main/goal/Goal.cs
@@ -6,7 +6,7 @@ using NutriApp.Workout;
 public interface Goal
 {
     double WeightGoal { get; }
-    int DailyCalorieGoal { get; }
+    double DailyCalorieGoal { get; }
     bool CheckWeight(double userWeight);
     void IncorporateFitness(List<Workout> recommendedWorkouts);
 }
